refactor: resolve todo
