+++ test/loader-generator-option.test.js
@@ -1,0 +1,320 @@
+import path from "path";
+import fileType from "file-type";
+import ImageMinimizerPlugin from "../src";
+
+import { runWebpack, fixturesPath } from "./helpers";
+
+describe("loader generator option", () => {
+  it("should work", async () => {
+    const outputDir = path.resolve(__dirname, "outputs", "loader-filename-1");
+    const stats = await runWebpack({
+      entry: path.join(fixturesPath, "./loader-single.js"),
+      imageminLoaderOptions: {
+        generator: [
+          {
+            preset: "webp",
+            implementation: ImageMinimizerPlugin.squooshGenerate,
+            options: {
+              encodeOptions: {
+                webp: {
+                  lossless: 1,
+                },
+              },
+            },
+          },
+        ],
+      },
+      output: {
+        path: outputDir,
+      },
+    });
+
+    const { compilation } = stats;
+    const { warnings, errors } = compilation;
+
+    const transformedAsset = path.resolve(
+      __dirname,
+      outputDir,
+      "loader-test.webp"
+    );
+
+    const transformedExt = await fileType.fromFile(transformedAsset);
+
+    expect(/image\/webp/i.test(transformedExt.mime)).toBe(true);
+    expect(warnings).toHaveLength(0);
+    expect(errors).toHaveLength(0);
+  });
+
+  it("should throw error on duplicate presets", async () => {
+    const outputDir = path.resolve(__dirname, "outputs", "loader-filename-1");
+    const stats = await runWebpack({
+      entry: path.join(fixturesPath, "./loader-single.js"),
+      imageminLoaderOptions: {
+        generator: [
+          {
+            preset: "webp",
+            implementation: ImageMinimizerPlugin.squooshGenerate,
+            options: {
+              encodeOptions: {
+                webp: {
+                  lossless: 1,
+                },
+              },
+            },
+          },
+          {
+            preset: "webp",
+            implementation: ImageMinimizerPlugin.squooshGenerate,
+            options: {
+              encodeOptions: {
+                webp: {
+                  lossless: 1,
+                },
+              },
+            },
+          },
+        ],
+      },
+      output: {
+        path: outputDir,
+      },
+    });
+
+    const { compilation } = stats;
+    const { warnings, errors } = compilation;
+
+    expect(warnings).toHaveLength(0);
+    expect(errors).toHaveLength(1);
+    expect(errors[0].message).toMatch(
+      /Found several identical pereset names, the 'preset' option should be unique/
+    );
+  });
+
+  it("should throw error on not found preset", async () => {
+    const outputDir = path.resolve(__dirname, "outputs", "loader-filename-1");
+    const stats = await runWebpack({
+      entry: path.join(fixturesPath, "./loader-single.js"),
+      imageminLoaderOptions: {
+        generator: [
+          {
+            preset: "webpz",
+            implementation: ImageMinimizerPlugin.squooshGenerate,
+            options: {
+              encodeOptions: {
+                webp: {
+                  lossless: 1,
+                },
+              },
+            },
+          },
+        ],
+      },
+      output: {
+        path: outputDir,
+      },
+    });
+
+    const { compilation } = stats;
+    const { warnings, errors } = compilation;
+
+    expect(warnings).toHaveLength(0);
+    expect(errors).toHaveLength(1);
+    expect(errors[0].message).toMatch(
+      /Can't find 'webp' preset in the 'generator' option/
+    );
+  });
+
+  it("should generate the new webp image with flat filename", async () => {
+    const outputDir = path.resolve(__dirname, "outputs", "loader-filename-1");
+    const stats = await runWebpack({
+      entry: path.join(fixturesPath, "./loader-single.js"),
+      imageminLoaderOptions: {
+        generator: [
+          {
+            preset: "webp",
+            implementation: ImageMinimizerPlugin.squooshGenerate,
+            filename: "[name].webp",
+            options: {
+              encodeOptions: {
+                webp: {
+                  lossless: 1,
+                },
+              },
+            },
+          },
+        ],
+      },
+      output: {
+        path: outputDir,
+      },
+    });
+
+    const { compilation } = stats;
+    const { warnings, errors } = compilation;
+
+    const transformedAsset = path.resolve(
+      __dirname,
+      outputDir,
+      "loader-test.webp"
+    );
+
+    const transformedExt = await fileType.fromFile(transformedAsset);
+
+    expect(/image\/webp/i.test(transformedExt.mime)).toBe(true);
+    expect(warnings).toHaveLength(0);
+    expect(errors).toHaveLength(0);
+  });
+
+  it("should generate the new webp image with nested filename", async () => {
+    const outputDir = path.resolve(__dirname, "outputs", "loader-filename-2");
+    const stats = await runWebpack({
+      entry: path.join(fixturesPath, "./loader-single.js"),
+      imageminLoaderOptions: {
+        generator: [
+          {
+            preset: "webp",
+            implementation: ImageMinimizerPlugin.squooshGenerate,
+            filename: "deep/[path][name].webp",
+            options: {
+              encodeOptions: {
+                webp: {
+                  lossless: 1,
+                },
+              },
+            },
+          },
+        ],
+      },
+      output: {
+        path: outputDir,
+      },
+    });
+    const { compilation } = stats;
+    const { warnings, errors } = compilation;
+
+    const transformedAsset = path.resolve(
+      __dirname,
+      outputDir,
+      "deep/nested/deep/loader-test.webp"
+    );
+
+    const transformedExt = await fileType.fromFile(transformedAsset);
+
+    expect(/image\/webp/i.test(transformedExt.mime)).toBe(true);
+    expect(warnings).toHaveLength(0);
+    expect(errors).toHaveLength(0);
+  });
+
+  it("should generate the new webp image with filename  pointing to other directory", async () => {
+    const outputDir = path.resolve(__dirname, "outputs", "loader-filename-3");
+    const stats = await runWebpack({
+      entry: path.join(fixturesPath, "./loader-single.js"),
+      imageminLoaderOptions: {
+        generator: [
+          {
+            preset: "webp",
+            implementation: ImageMinimizerPlugin.squooshGenerate,
+            filename: "other/[name].webp",
+            options: {
+              encodeOptions: {
+                webp: {
+                  lossless: 1,
+                },
+              },
+            },
+          },
+        ],
+      },
+      output: {
+        path: outputDir,
+      },
+    });
+    const { compilation } = stats;
+    const { warnings, errors } = compilation;
+
+    const transformedAsset = path.resolve(
+      __dirname,
+      outputDir,
+      "other/loader-test.webp"
+    );
+
+    const transformedExt = await fileType.fromFile(transformedAsset);
+
+    expect(/image\/webp/i.test(transformedExt.mime)).toBe(true);
+    expect(warnings).toHaveLength(0);
+    expect(errors).toHaveLength(0);
+  });
+
+  it("should generate the new webp image when filename is function", async () => {
+    const outputDir = path.resolve(__dirname, "outputs", "loader-filename-4");
+    const stats = await runWebpack({
+      entry: path.join(fixturesPath, "./loader-single.js"),
+      imageminLoaderOptions: {
+        generator: [
+          {
+            preset: "webp",
+            implementation: ImageMinimizerPlugin.squooshGenerate,
+            filename: () => "other/[name].webp",
+            options: {
+              encodeOptions: {
+                webp: {
+                  lossless: 1,
+                },
+              },
+            },
+          },
+        ],
+      },
+      output: {
+        path: outputDir,
+      },
+    });
+    const { compilation } = stats;
+    const { warnings, errors } = compilation;
+
+    const transformedAsset = path.resolve(
+      __dirname,
+      outputDir,
+      "other/loader-test.webp"
+    );
+
+    const transformedExt = await fileType.fromFile(transformedAsset);
+
+    expect(/image\/webp/i.test(transformedExt.mime)).toBe(true);
+    expect(warnings).toHaveLength(0);
+    expect(errors).toHaveLength(0);
+  });
+
+  it("should minimize image with flat filename", async () => {
+    const outputDir = path.resolve(__dirname, "outputs", "loader-filename-5");
+    const stats = await runWebpack({
+      entry: path.join(fixturesPath, "./simple.js"),
+      imageminLoaderOptions: {
+        minimizer: [
+          {
+            implementation: ImageMinimizerPlugin.squooshMinify,
+            filename: "[name][ext]",
+          },
+        ],
+      },
+      output: {
+        path: outputDir,
+      },
+    });
+
+    const { compilation } = stats;
+    const { warnings, errors } = compilation;
+
+    const transformedAsset = path.resolve(
+      __dirname,
+      outputDir,
+      "loader-test.jpg"
+    );
+
+    const transformedExt = await fileType.fromFile(transformedAsset);
+
+    expect(/image\/jpeg/i.test(transformedExt.mime)).toBe(true);
+    expect(warnings).toHaveLength(0);
+    expect(errors).toHaveLength(0);
+  });
+});
